refactor(video): Stop the texture size code confusing signed and unsigned
3 small helpers work out how large a texture needs to be: 2 round a
size up to the next power of 2 and 1 rounds it up to a multiple of the
hardware's tile size. All 3 took a signed number which advertises that
they can answer for negative sizes. They cannot. Asked to align -5 to a
multiple of four, the alignment helper returns zero; asked for the next
power of two of -5, the other two return 1. Neither is a rounded-up size.
They are simply whatever the arithmetic happened to produce.

A texture side is a count of pixels and is never negative, so the parameters
now say so, which removes those answers by making them unrepresentable
rather than by adding a check for a case that cannot occur. The size then
crosses between signed and unsigned in several further places in these two
files, because SDL states a surface's dimensions as signed and OpenGL takes
them back as signed while the texture stores them unsigned. Every one of
those crossings is now written out explicitly, including the ones that
predate this change, and the conversion from SDL's size happens once at the
top rather than at each use.

Making the helpers unsigned also removes work the compiler was doing on
their behalf. Signed division in C++ rounds toward zero, so the alignment
helper could not just mask off the low bits; it had to compare and
conditionally adjust first, purely to reproduce the behaviour for negatives
that we never wanted. These run once per texture load rather than in a hot
loop, so this is not a speed argument but more to do with code correctness.

+++ src/video/gl/gl_lightmap.cpp
@@ -44,9 +44,9 @@
 #include "video/surface.hpp"
 #include "video/texture_manager.hpp"
 
-inline int next_po2(int val)
+inline unsigned int next_po2(unsigned int val)
 {
-  int result = 1;
+  unsigned int result = 1;
   while(result < val)
     result *= 2;
 
@@ -62,8 +62,8 @@ GLLightmap::GLLightmap() :
 {
   m_lightmap_width = SCREEN_WIDTH / s_LIGHTMAP_DIV;
   m_lightmap_height = SCREEN_HEIGHT / s_LIGHTMAP_DIV;
-  unsigned int width = next_po2(m_lightmap_width);
-  unsigned int height = next_po2(m_lightmap_height);
+  unsigned int width = next_po2(static_cast<unsigned int>(m_lightmap_width));
+  unsigned int height = next_po2(static_cast<unsigned int>(m_lightmap_height));
 
   m_lightmap.reset(new GLTexture(width, height));
